+++ composer.json
@@ -9,11 +9,11 @@
     "type": "yii2-extension",
     "license": "BSD-3-Clause",
     "support": {
-        "issues": "https://github.com/yiisoft/yii2-collection/issues",
+        "issues": "https://github.com/KartaviK/yii2-collection/issues",
         "forum": "http://www.yiiframework.com/forum/",
         "wiki": "http://www.yiiframework.com/wiki/",
         "irc": "irc://irc.freenode.net/yii",
-        "source": "https://github.com/yiisoft/yii2-collection"
+        "source": "https://github.com/KartaviK/yii2-collection"
     },
     "authors": [
         {
@@ -30,25 +30,32 @@
         "yiisoft/yii2": "~2.0.14",
         "ext-json": "*"
     },
-    "repositories": [
-        {
-            "type": "composer",
-            "url": "https://asset-packagist.org"
-        }
-    ],
+    "require-dev": {
+        "phpunit/phpunit": "^7.5",
+        "squizlabs/php_codesniffer": "^3.4",
+        "vlucas/phpdotenv": "^2.5",
+        "yiisoft/yii2-phpunit": "dev-master"
+    },
     "autoload": {
         "psr-4": {
-            "Kartavik\\Yii\\Collection\\": "src"
+            "Kartavik\\Yii\\": "src"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Kartavik\\Yii\\Collection\\Tests\\": "tests"
+            "Kartavik\\Yii\\Tests\\": "tests"
         }
     },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0.x-dev"
+    "repositories": [
+        {
+            "type": "vcs",
+            "url": "git@github.com:Horat1us/yii2-phpunit.git"
+        }
+    ],
+    "config": {
+        "sort-packages": true,
+        "platform": {
+            "php": "7.1"
         }
     },
     "scripts": {
@@ -56,9 +63,5 @@
         "phpcbf": "phpcbf --standard=PSR2 ./src ./tests",
         "cover": "phpunit --coverage-clover=coverage.xml",
         "test": "phpunit"
-    },
-    "require-dev": {
-        "squizlabs/php_codesniffer": "^3.4",
-        "phpunit/phpunit": "^7.5"
     }
 }
