--- phpunit.xml.dist
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<phpunit bootstrap="./tests/bootstrap.php"
-		colors="true"
-		verbose="true"
-		stopOnFailure="false">
-		<testsuites>
-			<testsuite name="Test Suite">
-				<directory>./tests</directory>
-			</testsuite>
-		</testsuites>
-</phpunit>
