+++ tests/TestCase.php
@@ -5,66 +5,27 @@
  * @license http://www.yiiframework.com/license/
  */
 
-namespace yiiunit\collection;
+namespace Kartavik\Yii\Tests;
 
-use yii\db\Connection;
-use yii\helpers\ArrayHelper;
+use yii\phpunit;
+use yii\helpers;
 
-abstract class TestCase extends \PHPUnit\Framework\TestCase
+/**
+ * Class TestCase
+ * @package Kartavik\Yii\Tests
+ */
+abstract class TestCase extends phpunit\TestCase
 {
-    /**
-     * {@inheritdoc}
-     */
-    protected function setUp()
+    public function globalFixtures(): array
     {
-        $this->mockApplication();
-
-        \Yii::$app->db->createCommand()->createTable('customers', [
-            'id' => 'pk',
-            'name' => 'string NOT NULL',
-            'age' => 'integer NOT NULL',
-        ])->execute();
-
-        parent::setUp();
-    }
-
-    /**
-     * Clean up after test.
-     * By default the application created with [[mockApplication]] will be destroyed.
-     */
-    protected function tearDown()
-    {
-        parent::tearDown();
-        $this->destroyApplication();
-        \Yii::$container = new \yii\di\Container();
-    }
-
-    /**
-     * Populates Yii::$app with a new application
-     * The application will be destroyed on tearDown() automatically.
-     * @param array $config The application configuration, if needed
-     * @param string $appClass name of the application class to create
-     */
-    protected function mockApplication($config = [], $appClass = \yii\console\Application::class)
-    {
-        new $appClass(ArrayHelper::merge([
-            'id' => 'testapp',
-            'basePath' => __DIR__,
-            'vendorPath' => __DIR__ . '/../vendor',
-            'components' => [
-                'db' => [
-                    'class' => Connection::class,
-                    'dsn' => 'sqlite::memory:',
+        $fixtures = [
+            [
+                'class' => phpunit\MigrateFixture::class,
+                'migrationNamespaces' => [
+                    'Kartavik\\Yii\\Tests\\Migrations',
                 ],
-            ],
-        ], $config));
-    }
-
-    /**
-     * Destroys application in Yii::$app by setting it to null.
-     */
-    protected function destroyApplication()
-    {
-        \Yii::$app = null;
+            ]
+        ];
+        return helpers\ArrayHelper::merge(parent::globalFixtures(), $fixtures);
     }
 }
