--- tests/models/Customer.php
@@ -1,40 +1,0 @@
-<?php
-/**
- * @link http://www.yiiframework.com/
- * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
- */
-
-namespace yiiunit\collection\models;
-
-use yii\db\ActiveRecord;
-use yii\collection\CollectionBehavior;
-
-/**
- * Customer Model
- *
- * @property int $id
- * @property string $name
- * @property int $age
- */
-class Customer extends ActiveRecord
-{
-    /**
-     * {@inheritdoc}
-     */
-    public static function tableName()
-    {
-        return 'customers';
-    }
-
-    /**
-     * {@inheritdoc}
-     * @return \yii\db\ActiveQuery|CollectionBehavior
-     */
-    public static function find()
-    {
-        $query = parent::find();
-        $query->attachBehavior('collection', CollectionBehavior::class);
-        return $query;
-    }
-}
