--- tests/models/CustomerCollection.php
@@ -1,21 +1,0 @@
-<?php
-/**
- * @link http://www.yiiframework.com/
- * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
- */
-
-namespace yiiunit\collection\models;
-
-use yii\collection\ModelCollection;
-
-class CustomerCollection extends ModelCollection
-{
-    /**
-     * @return int
-     */
-    public function sumAge()
-    {
-        return $this->sum('age');
-    }
-}
